# src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "InvoiceReviewAssistant.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5190/"
    }
  ],
  "paths": {
...
    "/api/invoices/{id}/validate": {
      "post": {
        "tags": [
          "InvoiceValidation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateInvoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "validateInvoice"
      }
    }
  },
  "components": {
    "schemas": {
...
      "AmountFieldsDto": {
        "required": [
          "currency",
          "subtotal",
          "taxAmount",
          "total"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/TextFieldDto"
          },
          "subtotal": {
            "$ref": "#/components/schemas/MoneyFieldDto"
          },
          "taxAmount": {
            "$ref": "#/components/schemas/MoneyFieldDto"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyFieldDto"
          }
        },
        "additionalProperties": false
      },
...
      "ConfidenceBand": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low",
          "unknown"
        ]
      },
      "DateFieldDto": {
        "required": [
          "value",
          "originalValue",
          "originalSource",
          "currentSource",
          "confidence",
          "confidenceBand",
          "differsFromOriginal",
          "lastCorrectedAt"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "originalValue": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "originalSource": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "currentSource": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "confidence": {
            "type": [
              "null",
              "number"
            ],
            "format": "double",
            "minimum": 0,
            "maximum": 1
          },
          "confidenceBand": {
            "$ref": "#/components/schemas/ConfidenceBand"
          },
          "differsFromOriginal": {
            "type": "boolean"
          },
          "lastCorrectedAt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "DatesAndTermsFieldsDto": {
        "required": [
          "invoiceDate",
          "dueDate",
          "paymentTerms",
          "normalizedPaymentTermsDays"
        ],
        "type": "object",
        "properties": {
          "invoiceDate": {
            "$ref": "#/components/schemas/DateFieldDto"
          },
          "dueDate": {
            "$ref": "#/components/schemas/DateFieldDto"
          },
          "paymentTerms": {
            "$ref": "#/components/schemas/TextFieldDto"
          },
          "normalizedPaymentTermsDays": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DecisionKind": {
        "type": "string",
        "enum": [
          "approved",
          "rejected"
        ]
      },
      "DocumentIntegrityStatus": {
        "type": "string",
        "enum": [
          "available",
          "missing",
          "corrupt"
        ]
      },
      "DocumentTextSource": {
        "type": "string",
        "enum": [
          "nativeText",
          "ocr"
        ]
      },
      "FieldCorrectionDto": {
        "required": [
          "id",
          "auditEventId",
          "field",
          "previousValue",
          "newValue",
          "draftVersion",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^(?:0|[1-9]\\d*)$"
          },
          "auditEventId": {
            "type": "string",
            "pattern": "^(?:0|[1-9]\\d*)$"
          },
          "field": {
            "$ref": "#/components/schemas/InvoiceFieldKey"
          },
          "previousValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "newValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          }
        },
        "additionalProperties": false
      },
      "FieldSource": {
        "type": "string",
        "enum": [
          "nativeText",
          "ocr",
          "aiInference",
          "reviewer"
        ]
      },
      "InvoiceDecisionDto": {
        "required": [
          "kind",
          "decidedAt",
          "rejectionReason"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DecisionKind"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "rejectionReason": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "InvoiceDetailDto": {
        "required": [
          "id",
          "status",
          "draftVersion",
          "lastValidatedVersion",
          "createdAt",
          "updatedAt",
          "document",
          "documentTextSource",
          "fields",
          "reviewNotes",
          "summary",
          "currentValidation",
          "corrections",
          "processingFailure",
          "decision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "lastValidatedVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "document": {
            "$ref": "#/components/schemas/InvoiceDocumentDto"
          },
          "documentTextSource": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentTextSource"
              }
            ]
          },
          "fields": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InvoiceFieldsDto"
              }
            ]
          },
          "reviewNotes": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "$ref": "#/components/schemas/InvoiceSummaryDto"
          },
          "currentValidation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationRunDto"
              }
            ]
          },
          "corrections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldCorrectionDto"
            }
          },
          "processingFailure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProcessingFailureDto"
              }
            ]
          },
          "decision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InvoiceDecisionDto"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "InvoiceDocumentDto": {
        "required": [
          "originalFilename",
          "mediaType",
          "byteLength",
          "sha256",
          "pageCount",
          "integrityStatus"
        ],
        "type": "object",
        "properties": {
          "originalFilename": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "application/pdf"
            ]
          },
          "byteLength": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "integrityStatus": {
            "$ref": "#/components/schemas/DocumentIntegrityStatus"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceFieldKey": {
        "type": "string",
        "enum": [
          "supplierName",
          "supplierRegistrationId",
          "invoiceNumber",
          "purchaseOrderNumber",
          "invoiceDate",
          "dueDate",
          "paymentTerms",
          "currency",
          "subtotal",
          "taxAmount",
          "total",
          "reviewNotes"
        ]
      },
      "InvoiceFieldsDto": {
        "required": [
          "supplier",
          "reference",
          "datesAndTerms",
          "amounts"
        ],
        "type": "object",
        "properties": {
          "supplier": {
            "$ref": "#/components/schemas/SupplierFieldsDto"
          },
          "reference": {
            "$ref": "#/components/schemas/ReferenceFieldsDto"
          },
          "datesAndTerms": {
            "$ref": "#/components/schemas/DatesAndTermsFieldsDto"
          },
          "amounts": {
            "$ref": "#/components/schemas/AmountFieldsDto"
          }
        },
        "additionalProperties": false
      },
      "InvoiceProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 100
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "fields": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            ]
          },
          "currentVersion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 1
              }
            ]
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "fields",
          "currentVersion"
        ],
        "additionalProperties": false
      },
...
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "processing",
          "reviewRequired",
          "readyForApproval",
          "approved",
          "rejected",
          "processingFailed"
        ]
      },
      "InvoiceSummaryDto": {
        "required": [
          "extractedFieldCount",
          "warningCount",
          "errorCount",
          "manualCorrectionCount"
        ],
        "type": "object",
        "properties": {
          "extractedFieldCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "warningCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "manualCorrectionCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "MoneyFieldDto": {
        "required": [
          "value",
          "originalValue",
          "originalSource",
          "currentSource",
          "confidence",
          "confidenceBand",
          "differsFromOriginal",
          "lastCorrectedAt"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "string"
            ],
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "originalValue": {
            "type": [
              "null",
              "string"
            ],
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "originalSource": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "currentSource": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "confidence": {
            "type": [
              "null",
              "number"
            ],
            "format": "double",
            "minimum": 0,
            "maximum": 1
          },
          "confidenceBand": {
            "$ref": "#/components/schemas/ConfidenceBand"
          },
          "differsFromOriginal": {
            "type": "boolean"
          },
          "lastCorrectedAt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "ProcessingFailureCode": {
        "type": "string",
        "enum": [
          "PDF_EXTRACTION_FAILED",
          "PDF_RENDER_FAILED",
          "OCR_UNAVAILABLE",
          "OCR_PAGE_TIMEOUT",
          "OCR_DOCUMENT_TIMEOUT",
          "OCR_FAILED",
          "AI_TIMEOUT",
          "AI_UNAVAILABLE",
          "AI_REFUSED",
          "AI_RESPONSE_INCOMPLETE",
          "AI_RESPONSE_INVALID",
          "PROCESS_INTERRUPTED",
          "PROCESSING_FAILED"
        ]
      },
      "ProcessingFailureDto": {
        "required": [
          "stage",
          "code",
          "message",
          "failedAt"
        ],
        "type": "object",
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/ProcessingStage"
          },
          "code": {
            "$ref": "#/components/schemas/ProcessingFailureCode"
          },
          "message": {
            "type": "string"
          },
          "failedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          }
        },
        "additionalProperties": false
      },
      "ProcessingStage": {
        "type": "string",
        "enum": [
          "upload",
          "pdfExtraction",
          "ocr",
          "aiExtraction",
          "parsing",
          "persistence",
          "startupRecovery"
        ]
      },
...
      "ReferenceFieldsDto": {
        "required": [
          "invoiceNumber",
          "purchaseOrderNumber"
        ],
        "type": "object",
        "properties": {
          "invoiceNumber": {
            "$ref": "#/components/schemas/TextFieldDto"
          },
          "purchaseOrderNumber": {
            "$ref": "#/components/schemas/TextFieldDto"
          }
        },
        "additionalProperties": false
      },
...
      "SupplierFieldsDto": {
        "required": [
          "name",
          "registrationId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/TextFieldDto"
          },
          "registrationId": {
            "$ref": "#/components/schemas/TextFieldDto"
          }
        },
        "additionalProperties": false
      },
      "TextFieldDto": {
        "required": [
          "value",
          "originalValue",
          "originalSource",
          "currentSource",
          "confidence",
          "confidenceBand",
          "differsFromOriginal",
          "lastCorrectedAt"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "originalValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "originalSource": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "currentSource": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "confidence": {
            "type": [
              "null",
              "number"
            ],
            "format": "double",
            "minimum": 0,
            "maximum": 1
          },
          "confidenceBand": {
            "$ref": "#/components/schemas/ConfidenceBand"
          },
          "differsFromOriginal": {
            "type": "boolean"
          },
          "lastCorrectedAt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "ValidateInvoiceRequest": {
        "required": [
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ValidationCode": {
        "type": "string",
        "enum": [
          "REQUIRED_FIELD_MISSING",
          "AMOUNT_RECONCILIATION_FAILED",
          "NEGATIVE_AMOUNT_UNEXPECTED",
          "DUE_DATE_BEFORE_INVOICE_DATE",
          "PAYMENT_TERMS_MISMATCH",
          "POSSIBLE_DUPLICATE_INVOICE",
          "CURRENCY_INVALID",
          "LOW_EXTRACTION_CONFIDENCE",
          "INVOICE_DATE_IN_FUTURE"
        ]
      },
      "ValidationResultDto": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ValidationCode"
          },
          "severity": {
            "$ref": "#/components/schemas/ValidationSeverity"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceFieldKey"
            }
          },
          "data": {}
        },
        "required": [
          "code",
          "severity",
          "message",
          "fields",
          "data"
        ],
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "code",
          "mapping": {
            "REQUIRED_FIELD_MISSING": "#/components/schemas/RequiredFieldMissingValidationResultDto",
            "AMOUNT_RECONCILIATION_FAILED": "#/components/schemas/AmountReconciliationFailedValidationResultDto",
            "NEGATIVE_AMOUNT_UNEXPECTED": "#/components/schemas/NegativeAmountUnexpectedValidationResultDto",
            "DUE_DATE_BEFORE_INVOICE_DATE": "#/components/schemas/DueDateBeforeInvoiceDateValidationResultDto",
            "PAYMENT_TERMS_MISMATCH": "#/components/schemas/PaymentTermsMismatchValidationResultDto",
            "POSSIBLE_DUPLICATE_INVOICE": "#/components/schemas/PossibleDuplicateInvoiceValidationResultDto",
            "CURRENCY_INVALID": "#/components/schemas/CurrencyInvalidValidationResultDto",
            "LOW_EXTRACTION_CONFIDENCE": "#/components/schemas/LowExtractionConfidenceValidationResultDto",
            "INVOICE_DATE_IN_FUTURE": "#/components/schemas/InvoiceDateInFutureValidationResultDto"
          }
        }
      },
      "ValidationRunDto": {
        "required": [
          "id",
          "draftVersion",
          "validatedAt",
          "warningCount",
          "errorCount",
          "results"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "validatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationResultDto"
            }
          }
        },
        "additionalProperties": false
      },
      "ValidationSeverity": {
        "type": "string",
        "enum": [
          "warning",
          "error"
        ]
      },
...
      "RequiredFieldMissingValidationDataDto": {
        "type": "object",
        "properties": {
          "missingField": {
            "$ref": "#/components/schemas/InvoiceFieldKey"
          }
        },
        "required": [
          "missingField"
        ],
        "additionalProperties": false
      },
      "AmountReconciliationFailedValidationDataDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string"
          },
          "subtotal": {
            "type": "string",
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "taxAmount": {
            "type": "string",
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "expectedTotal": {
            "type": "string",
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "actualTotal": {
            "type": "string",
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "difference": {
            "type": "string",
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "tolerance": {
            "type": "string",
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          }
        },
        "required": [
          "currency",
          "subtotal",
          "taxAmount",
          "expectedTotal",
          "actualTotal",
          "difference",
          "tolerance"
        ],
        "additionalProperties": false
      },
...
      "RequiredFieldMissingValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/RequiredFieldMissingValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "AmountReconciliationFailedValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/AmountReconciliationFailedValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "NegativeAmountUnexpectedValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/NegativeAmountUnexpectedValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "DueDateBeforeInvoiceDateValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DueDateBeforeInvoiceDateValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "PaymentTermsMismatchValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PaymentTermsMismatchValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "PossibleDuplicateInvoiceValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PossibleDuplicateInvoiceValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "CurrencyInvalidValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CurrencyInvalidValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "LowExtractionConfidenceValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/LowExtractionConfidenceValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "InvoiceDateInFutureValidationResultDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ValidationResultDto"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/InvoiceDateInFutureValidationDataDto"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
...
    }
  },
...
}